Separated merge_exposures into a separate file from main.rs
